// src/api/openapi/auto-generated/models/CreateSingleLocationReportInput.ts
export interface CreateSingleLocationReportInput {
    /**
     * 
     * @type {SingleLocationReportType}
     * @memberof CreateSingleLocationReportInput
     */
    type?: SingleLocationReportType | null;
    /**
     * 
     * @type {Array<number>}
     * @memberof CreateSingleLocationReportInput
     */
    locationIds?: Array<number> | null;
    /**
     * 
     * @type {string}
     * @memberof CreateSingleLocationReportInput
     */
    templateId?: string | null;
    /**
     * 
     * @type {JupiterTemplateInput}
     * @memberof CreateSingleLocationReportInput
     */
    jupiterTemplate: JupiterTemplateInput;
    /**
     * 
     * @type {string}
     * @memberof CreateSingleLocationReportInput
     */
    perilsResultSetId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateSingleLocationReportInput
     */
    impactsResultSetId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateSingleLocationReportInput
     */
    scoresResultSetId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateSingleLocationReportInput
     */
    disclosureResultSetId?: string | null;
}