// src/api/openapi/auto-generated/models/DamagesResultSetOptions.ts
export interface DamagesResultSetOptions {
    /**
     * 
     * @type {string}
     * @memberof DamagesResultSetOptions
     */
    eiVersion?: DamagesResultSetOptionsEiVersionEnum;
    /**
     * 
     * @type {PerilsOptions}
     * @memberof DamagesResultSetOptions
     */
    perilsOptions?: PerilsOptions;
    /**
     * 
     * @type {Array<APIEIModule>}
     * @memberof DamagesResultSetOptions
     */
    eiModules?: Array<APIEIModule>;
    /**
     * 
     * @type {boolean}
     * @memberof DamagesResultSetOptions
     */
    windPeakGustAdjustment?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof DamagesResultSetOptions
     */
    includePeakGustFactor?: boolean;
    /**
     * 
     * @type {LossAssetAttributes}
     * @memberof DamagesResultSetOptions
     */
    defaults?: LossAssetAttributes;
    /**
     * 
     * @type {string}
     * @memberof DamagesResultSetOptions
     */
    type: DamagesResultSetOptionsTypeEnum;
}


/**
 * @export
 */
export const DamagesResultSetOptionsEiVersionEnum = {
    _262: '2.6.2',
    _300: '3.0.0',
    _310: '3.1.0',
    _311: '3.1.1',
    _312: '3.1.2',
    _320: '3.2.0',
    _330: '3.3.0',
    unknownDefaultOpenApi: '11184809'
} as const;
export type DamagesResultSetOptionsEiVersionEnum = typeof DamagesResultSetOptionsEiVersionEnum[keyof typeof DamagesResultSetOptionsEiVersionEnum];

/**
 * @export
 */
export const DamagesResultSetOptionsTypeEnum = {
    damages: 'damages',
    unknownDefaultOpenApi: '11184809'
} as const;
export type DamagesResultSetOptionsTypeEnum = typeof DamagesResultSetOptionsTypeEnum[keyof typeof DamagesResultSetOptionsTypeEnum];