// src/api/openapi/auto-generated/models/JupiterTemplateColorCoding.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const JupiterTemplateColorCoding = {
    exposure: 'exposure',
    change: 'change',
    unknownDefaultOpenApi: '11184809'
} as const;
export type JupiterTemplateColorCoding = typeof JupiterTemplateColorCoding[keyof typeof JupiterTemplateColorCoding];


export function instanceOfJupiterTemplateColorCoding(value: any): boolean {
    for (const key in JupiterTemplateColorCoding) {
        if (Object.prototype.hasOwnProperty.call(JupiterTemplateColorCoding, key)) {
            if ((JupiterTemplateColorCoding as Record<string, JupiterTemplateColorCoding>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function JupiterTemplateColorCodingFromJSON(json: any): JupiterTemplateColorCoding {
    return JupiterTemplateColorCodingFromJSONTyped(json, false);
}

export function JupiterTemplateColorCodingFromJSONTyped(json: any, ignoreDiscriminator: boolean): JupiterTemplateColorCoding {
    return json as JupiterTemplateColorCoding;
}

export function JupiterTemplateColorCodingToJSON(value?: JupiterTemplateColorCoding | null): any {
    return value as any;
}