// src/api/openapi/auto-generated/models/JupiterTemplateScenario.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const JupiterTemplateScenario = {
    ssp2: 'ssp2',
    ssp5: 'ssp5',
    unknownDefaultOpenApi: '11184809'
} as const;
export type JupiterTemplateScenario = typeof JupiterTemplateScenario[keyof typeof JupiterTemplateScenario];


export function instanceOfJupiterTemplateScenario(value: any): boolean {
    for (const key in JupiterTemplateScenario) {
        if (Object.prototype.hasOwnProperty.call(JupiterTemplateScenario, key)) {
            if ((JupiterTemplateScenario as Record<string, JupiterTemplateScenario>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function JupiterTemplateScenarioFromJSON(json: any): JupiterTemplateScenario {
    return JupiterTemplateScenarioFromJSONTyped(json, false);
}

export function JupiterTemplateScenarioFromJSONTyped(json: any, ignoreDiscriminator: boolean): JupiterTemplateScenario {
    return json as JupiterTemplateScenario;
}

export function JupiterTemplateScenarioToJSON(value?: JupiterTemplateScenario | null): any {
    return value as any;
}