// src/api/openapi/auto-generated/models/JupiterTemplateType.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const JupiterTemplateType = {
    basic: 'basic',
    standard: 'standard',
    advanced: 'advanced',
    fidelity: 'fidelity',
    csrd: 'csrd',
    unknownDefaultOpenApi: '11184809'
} as const;
export type JupiterTemplateType = typeof JupiterTemplateType[keyof typeof JupiterTemplateType];


export function instanceOfJupiterTemplateType(value: any): boolean {
    for (const key in JupiterTemplateType) {
        if (Object.prototype.hasOwnProperty.call(JupiterTemplateType, key)) {
            if ((JupiterTemplateType as Record<string, JupiterTemplateType>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function JupiterTemplateTypeFromJSON(json: any): JupiterTemplateType {
    return JupiterTemplateTypeFromJSONTyped(json, false);
}

export function JupiterTemplateTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): JupiterTemplateType {
    return json as JupiterTemplateType;
}

export function JupiterTemplateTypeToJSON(value?: JupiterTemplateType | null): any {
    return value as any;
}