// src/api/openapi/auto-generated/models/SingleLocationReportType.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const SingleLocationReportType = {
    singleLocation: 'single_location',
    unknownDefaultOpenApi: '11184809'
} as const;
export type SingleLocationReportType = typeof SingleLocationReportType[keyof typeof SingleLocationReportType];


export function instanceOfSingleLocationReportType(value: any): boolean {
    for (const key in SingleLocationReportType) {
        if (Object.prototype.hasOwnProperty.call(SingleLocationReportType, key)) {
            if ((SingleLocationReportType as Record<string, SingleLocationReportType>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SingleLocationReportTypeFromJSON(json: any): SingleLocationReportType {
    return SingleLocationReportTypeFromJSONTyped(json, false);
}

export function SingleLocationReportTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): SingleLocationReportType {
    return json as SingleLocationReportType;
}

export function SingleLocationReportTypeToJSON(value?: SingleLocationReportType | null): any {
    return value as any;
}